% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpa-narratives.R
\docType{data}
\name{wpa_narratives}
\alias{wpa_narratives}
\title{WPA former slave narratives}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2259 rows and 12 columns.}
\usage{
wpa_narratives
}
\description{
WPA former slave narratives
}
\keyword{datasets}

